pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.1",
    "langchain-cohere>=0.5.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=0.3.0",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.52.2",
]
